Added an implementation of disjoint set
I have 2 disjoint set implementation. 1 likely has bugs and need to be
removed. The one inside of datastruct::mp is usable. but needs testing.
Added a new test case that attempts to reconstruct the 5 images in a
bruteforce manner. (work in progress).

Additionally added printinfo functions (with varying verbosity) for
image feaures and pair matcher.

Added comparison operators for featureIDs.